requirements.txt
Django>=2.1.3,<=2.2.0
djangorestframework>=3.9.0,<=3.10.0
psycopg2>=2.8.3,<=2.8.3
PilloW>=6.1.0,<=6.1.0
flake8>=3.7.8,<=3.7.8